requirements.txt
streamlit
ibm-watson
python-dotenv
